ml-api/requirements.txt
numpy<2
transformers==4.35.2
fastapi
uvicorn
python-dotenv
pillow
python-multipart
aiofiles
google-generativeai
